Add confirmation delete messages

+++ public/js/app.js
@@ -377,7 +377,7 @@ function addTeam() {
           var proceed = confirm("The Current Team contains " + cards.length +" cards on it, applying this changes will remove every single card from it, Do you want to proceed ?");
           if(proceed) {
             $(cards).each(function(index, item){
-                $(item).appendTo("#peopleDiv");
+              $(item).appendTo("#peopleDiv");
             });
             $('#' + team_id).html(teaming.renderTeamTemplate(team, team_id));
           }
@@ -449,14 +449,20 @@ function restoreState(id) {
 }
 
 function deleteAll() {
-  localStorage.clear();
-  location.reload();
+  var proceed = confirm("Are you sure to delete all scenarios?");
+  if(proceed) {
+    localStorage.clear();
+    location.reload();
+  }
 }
 
 function deleteState(id) {
-  $("#" + id).remove();
-  localStorage.removeItem(id);
-  saveScenariosContainerState();
+  var proceed = confirm("Are you sure to delete this item?");
+  if(proceed) {
+    $("#" + id).remove();
+    localStorage.removeItem(id);
+    saveScenariosContainerState();
+  }
 }
 
 function saveScenariosContainerState() {
@@ -472,10 +478,20 @@ function guid() {
 }
 
 function deleteTeam(id){
+  var proceed = confirm("Are you sure to delete this item?");
+  if(proceed) {
+    var cards = $('#' + id).find("div.card");
+    $(cards).each(function(index, item){
+      $(item).appendTo("#peopleDiv");
+    });
+
     $("#" + id).remove();
+  }
 }
 
 function deletePerson(id){
+  var proceed = confirm("Are you sure to delete this item?");
+  if(proceed) {
     $("#" + id).remove();
+  }
 }
-
